move on_device op to annotation namespace
