+++ include/tvm/relay/attrs/device_copy.h
@@ -1,7 +1,7 @@
 /*!
  *  Copyright (c) 2018 by Contributors
  * \file tvm/relay/attrs/device_copy.h
- * \brief Attribute for device copy operator.
+ * \brief Attribute for the device copy operator.
  */
 #ifndef TVM_RELAY_ATTRS_DEVICE_COPY_H_
 #define TVM_RELAY_ATTRS_DEVICE_COPY_H_
@@ -11,20 +11,6 @@
 
 namespace tvm {
 namespace relay {
-
-/*!
- * \brief Options for the device annotation operators.
- */
-struct OnDeviceAttrs : public tvm::AttrsNode<OnDeviceAttrs> {
-  int device_type;
-
-  TVM_DECLARE_ATTRS(OnDeviceAttrs, "relay.attrs.OnDeviceAttrs") {
-    TVM_ATTR_FIELD(device_type)
-      .describe(
-         "The virutal device/context type that an expression is annotated with.")
-      .set_default(0);
-  }
-};
 
 /*!
  * \brief Options for the device copy operators.
