+++ python/tvm/relay/op/annotation/__init__.py
@@ -1,0 +1,4 @@
+# pylint: disable=wildcard-import
+"""Annotation related operators."""
+from __future__ import absolute_import as _abs
+from .annotation import *
