+++ python/tvm/relay/op/annotation/_make.py
@@ -1,0 +1,4 @@
+"""Constructor APIs"""
+from ...._ffi.function import _init_api
+
+_init_api("relay.op.annotation._make", __name__)
