+++ src/relay/op/device_copy.cc
@@ -45,27 +45,5 @@ on different devices.
 .set_attr<FInferCorrectLayout>("FInferCorrectLayout",
                                ElemwiseArbitraryLayout);
 
-// relay.on_device
-TVM_REGISTER_NODE_TYPE(OnDeviceAttrs);
-
-TVM_REGISTER_API("relay.op._make.on_device")
-.set_body_typed<Expr(Expr, int)>([](Expr data, int device_type) {
-  auto attrs = make_node<OnDeviceAttrs>();
-  attrs->device_type = device_type;
-  static const Op& op = Op::Get("on_device");
-  return CallNode::make(op, {data}, Attrs(attrs), {});
-});
-
-RELAY_REGISTER_OP("on_device")
-.describe(R"code(Annotate an expression with device type)code" TVM_ADD_FILELINE)
-.set_num_inputs(1)
-.set_support_level(10)
-.add_type_rel("Identity", IdentityRel)
-.set_attr<TOpPattern>("TOpPattern", kOpaque)
-.set_attr<TOpIsStateful>("TOpIsStateful", false)
-.set_attr<FInferCorrectLayout>("FInferCorrectLayout",
-                               ElemwiseArbitraryLayout);
-
-
 }  // namespace relay
 }  // namespace tvm
